requirements.txt
# Core framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
python-dotenv==1.0.1

# A2A Protocol & Streaming
sse-starlette==2.2.1
httpx==0.28.1
httpx-sse==0.4.0

# AWS Bedrock for Claude 4.5
boto3==1.35.81
botocore==1.35.81

# LangGraph & LangChain
langgraph==0.2.59
langchain==0.3.13
langchain-aws==0.2.10
langchain-core==0.3.28

# Redis for state management
redis==5.2.1
hiredis==3.1.0

# Utilities
uuid6==2024.7.10
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-httpx==0.34.0

# Type checking
mypy==1.13.0
types-redis==4.6.0.20241004

